refactoring file for MVC
- refactoring model file
- checkID is no longer need it, ,mongodb take care of it
- later, implement controller func again
