+++ routes/tourRoutes.js
@@ -9,7 +9,7 @@ const router = express.Router(); // create new route and save to var
 
 // PARAM MIDDLEWARE -> only run for certain parameters in URL
 // only work for this route (tour)
-router.param('id', tourController.checkID);
+// router.param('id', tourController.checkID);
 
 // ROUTER
 router
