+++ server.js
@@ -8,7 +8,7 @@ const app = require('./app.js');
 
 dotenv.config({ patch: './config.env' });
 
-// CONNECT TO DB
+// == CONNECT TO DB
 const DB = process.env.DATABASE.replace(
   '<PASSWORD>',
   process.env.DATABASE_PASSWORD
@@ -24,44 +24,11 @@ mongoose
     console.log('DB connection successfully');
   });
 
-// SCHEMA: describe data and validator
-const tourSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, 'A tour must have a name'],
-    unique: true,
-  },
-  rating: {
-    type: Number,
-    default: 4.5,
-  },
-  price: {
-    type: Number,
-    required: [true, 'A tour must have a price'],
-  },
-});
-
-// MODEL -> name of model and schema
-const Tour = mongoose.model('Tour', tourSchema); // create tour out or schema
-
-// DOCUMENT
-const testTour = new Tour({
-  name: 'The testtt',
-});
-testTour
-  .save() // reture promise
-  .then((doc) => {
-    console.log(doc);
-  })
-  .catch((err) => {
-    console.log('ERROR !!!!!: ', err);
-  });
-
-// ENVIRONMENT
+// == ENVIRONMENT
 console.log(`Environment: ${app.get('env')}`);
 // console.log(process.env);
 
-// START SERVER
+// == START SERVER
 const port = process.env.PORT || 3000;
 // listening to req
 app.listen(port, () => {
